Merged PR 8826: IVI-3056: Run and publish tests for the example app
For now, run and publish unit tests. The results are visible under the "Tests" tab.
Prepared an extra job to handle instrumented tests in the near future.
Example run here:

https://dev.azure.com/tomtomweb/IVI-Developers/_build/results?buildId=275395&view=results

T.B.D. whether or not we want to expose navtest and the emulators plugin.
T.B.D. if we want to publish any pipeline artifacts. If so what should be published.