Merged PR 10079: IVI-3280: Create the SDK release pipeline
Initial version.

Currently it contains two stages:
- A publishing stage to publish artifacts to Nexus.
- A verification stage to build the example app using dependencies from the IVI Nexus repository.

When navigating to this pipeline using the DevOps UI, the user can enter a version number per artifact so that this pipeline knows which files it needs to publish to Nexus. Later on we could also choose to programmatically always get the latest version available, before publishing it to the IVI Nexus repository.

Other verification jobs will follow.

[Example run](https://dev.azure.com/tomtomweb/IVI-Developers/_build/results?buildId=347079&view=results)