Merged PR 11413: IVI-3840 Use static and dynamic configurations
IVI-3840 Use static and dynamic configurations

The account service uses a static configuration to obtain URI for online
login. It does not use URI tho.

The accout settings service uses a dynamic configuration to obtain the
number of days the account may stay logged in. It is exposed as
reasd-only setting. Then the account service uses the setting to
validate status of the active account.