Merged PR 11539: IVI-3961: Generate HTML documentation
Added a Gradle task (`./gradlew docs`) to the IVI Example app to generate HTML documents from the available Markdown files.

Dokka was not suitable for this task so I used a simple command-line tool ([pandoc](https://pandoc.org/)) and a shell script to execute the conversion.

Added the _'Getting Started'_ document to the portal folder as well for ease of conversion.

If pandoc is not installed, the script exits with a warning to install pandoc before running the Gradle task.