Merged PR 12087: IVI-3159: Developer "setting up" documentation (Windows)
I tested the steps I added to the document on a Windows Enterprise VM.
I was able to build the Example App taking these steps, and configuring the IndiGO Emulator.

Unfortunately, I was not able to actually run the IndiGO Emulator, because of limitations on the Windows VM I'm using. It was not possible to install the `Intel x86 Emulator Accelerator (HAXM installer)` on the VM:

**Required SDK Tool**
![Screenshot from 2021-06-30 10-31-51.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/12087/attachments/Screenshot%20from%202021-06-30%2010-31-51.png)

**Installer fails**
![Screenshot from 2021-06-30 10-32-14.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/12087/attachments/Screenshot%20from%202021-06-30%2010-32-14.png)

**Installation log**
![Screenshot from 2021-06-30 10-32-56.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/12087/attachments/Screenshot%20from%202021-06-30%2010-32-56.png)

I tried configuring VirtualBox to support [EPT](https://en.wikipedia.org/wiki/Second_Level_Address_Translation#Extended_Page_Tables) but I didn't succeed with that.
See also [this thread](https://forums.virtualbox.org/viewtopic.php?f=6&t=101920) on VirtualBox forum.

For verification of this ticket, we have discussed @<Alexis Riksman> will try to run the Example App on the IndiGO Emulator on his Windows machine and see if everything works like intended.