Merged PR 13313: IVI-4637: Add HTML documentation to folder
HTML documentation can be generated running `./gradlew docs` and will now be saved to `/docs/portal/html`.
Also fixed broken links in the documentation (from `.md` files to `.html`).

Instead of adding the generated HTML documentation to the Example App repo, it should be generated as part of the SDK Release script.
This will be picked up in a different PR: [placeholder ticket].