Merged PR 14174: IVI 4915: Update the portal files to the site-map 'essentials'
Split Getting Started into multiple pages.
Updated the anchorlinks within the pages.
Added the images that relate to Getting Started into its own `/images` folder.
Added the files for the API reference to `/src`.
Moved "Installing on a custom device" page to FAQ.

The file structure for all essentials is now in place.