Merged PR 15628: IVI-5296: Add an extra stage to the IndiGO SDK release pipeline to generate and publish the API docs
In order for the indigo API reference documentation to work properly (with links to the tomtom-android-tools api documentation), we need to generate the tomtom-android-tools api documentation and make it available on the server before we generate the IndiGO api documentation.
This means that we should not use the indigo api documentation available on artifactory.

Added a stage to run in parallel to the Nexus publishing stage (to save time) to generate the api docs and push it to the S3 server (under a <version> folder):

```
/documents # aws s3 ls s3://devportal-bucket/downloads/indigo/indigo-api/
                           PRE 1.0.2556/
                           PRE latest/
/documents # aws s3 ls s3://devportal-bucket/downloads/indigo/android-tools-api/
                           PRE 0.0.223/
                           PRE latest/
```

Only sync that specific version of the api documentation to the **latest** folder, if verification of the SDK release succeeds.

Example:
![image (3).png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/15628/attachments/image%20%283%29.png)