Merged PR 16354: DEV: Use a Nexus account with restricted access to build the example app.
Alexa binaries on Nexus are hidden in a special group, that only some customers have access to.
The current account that we use in the SDK release pipeline has full access to the ivi repostory on Nexus.
Use a different account to **_assemble_** the example-app, that does not have access to the Alexa binaries.
![image.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/16354/attachments/image.png)

See [slack conversation](https://tomtomslack.slack.com/archives/CQQ9ZQ0F9/p1636120694028400)