Merged PR 16498: IVI-5623: Generate and upload Comms SDK API docs from the IndiGO SDK Release pipeline.
Generate and upload the Communications SDK API documentation, before generating the IndiGO API documentations

Note: We will have to start using an IndiGO version that is using at least v**0.1.76** of the communications SDK to be able to generate the comms sdk api docs. Otherwise the `./gradlew docs` task will give an `401:  Unauthorized error` because of outdated hardcoded credentials that were being used in the repositories file.