Merged PR 17178: IVI-5701: Use the correct artifacts to run the intake pipeline from a PR
This change allows a developer to run the **IndiGO Intake into IndiGO Docs** pipeline from a PR created for the **indigo** repository.

Added a job to generate & publish the libraries  to the **ivi-test** repository on Artifactory, based on the developer's changes made in his forked **indigo** repository. Because otherwise we first need to run the **required** (indigo) build and wait for it to publish the libraries after 30+ minutes, before a **IndiGO Intake into IndiGO Docs** run can be started . So let's take a shortcut instead. This job only runs in case the pipeline is triggered from a PR.

The version of the artifacts in the above situation is derived from the branch-name and commit hash.
Renamed the canary branch-name to make it unique for each commit added to the PR, to avoid conflicts.

The rest of the intake pipeline should run as usual, except no merge takes place at the end in case the pipeline was triggered from a PR.

**Note**: to make the intake pipeline appear in a PR. We will still need to set it up in the **Project settings** on Azure

Examples:
[ivi-test repo on Artifactory](https://artifactory.navkit-pipeline.tt3.com/artifactory/ivi-test/com/tomtom/ivi/platform/)
[Example PR (in IVI-Developers)](https://dev.azure.com/tomtomweb/IVI-Developers/_git/indigo.diogo.bilkes/pullrequest/17504)
[Example PR run (in IVI-Developers)](https://dev.azure.com/tomtomweb/IVI-Developers/_build/results?buildId=1111424&view=logs&j=d983a281-8055-5423-069b-0bd0c8787f91&t=18ae67dd-e536-581c-ecac-609f76224c10)

![image.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/17178/attachments/image.png)