Merged PR 18716: IVI-6192: Publish SDK release in the very last "Publication" stage.
Previously the Documentation or Verification stage could fail. But the tarball could have been published to Nexus already.

Changed the logic to only publish the tomtom-indigo-sdk tarball to Nexus in the very last stage when everything else passes.
For the verification stage the release content is already prepared upfront, but published as a pipeline artifact. Use the pipeline artifact to run tests with.

Moved all api-reference jobs to a separate template to keep the main yaml file more readable