Merged PR 18757: IVI-5905: Update the validation script + pipeline task
When running the `portal_check` Gradle task, API links are now generated from API Reference index files hosted on Artifactory instead of the API References hosted on the S3 bucket, removing the dependency on a SDK release for updating API symbols in the documentation.

Because of this, the API Reference URLs generated during the `portal_check` task might not (yet) be online (hosted on S3) and are exempt from URL validation.

Because the dependency on a SDK release has been removed, the `portal_check` Gradle task can now fail in the IndiGO Docs pipeline, instead of its previous behavior of only producing warnings. This means developers are required to fix issues in Documentation for their PRs to pass.

The changed logic now [enables us to use the Versions.toml file directly for the portal-related scripts](https://jira.tomtomgroup.com/browse/IVI-5993), instead of having its own versioning.

This PR also fixes the removal of the formerly used JSON file, and indexes the downloaded API References to create a lookup map.