Merged PR 19916: DEV: Integrate IndiGO v1.0.3309
IndiGO increased the dependency of `kotlin-tools` which uses a newer coroutines version than what IndiGO is compatible with. This isn't an issue in IndiGO as the coroutines version is forced to a specific (older) one. As a proper solution:
* IndiGO should provide a BOM file that customers can use to automatically adopt the right IndiGO versions as well. (It's on Archipelago's backlog.)
* IndiGO should update its coroutines version. (On Undergo's backlog.)

Until then however, each product must force the right version like IndiGO does. This is necessary because even if we'd revert the `kotlin-tools` version, any other library update could cause the same issue.