Merged PR 20338: IVI-6823 Remove references to moved classes in guide
`MediaSourcePanel`, `MediaSourceViewModel` and `MediaSourceFragment` are no longer available in the public API. Whilst the `MediaTaskPanel` is available, I decided not to reference it here, since it's already referenced at the "Example panel" section of this very same guide.

Important to fix this quickly, since this prevents the portal from being published.