Merged PR 22411: IVI-7315 Create custom panel type example
This is still pretty rough, need an obvious pass for polishing. Especially docs are still missing. It _does_ already gives a good idea whether the solution is in the right direction, though.

The example shows the implementation of a `CustomSystemUiPanel`. It includes an interaction between frontend state and different panels, in this case whether the `ExampleFrontend` is open.

This is what it looks like:
![screencap.png](https://dev.azure.com/tomtomweb/815cc923-fd52-415e-b765-bc4a05b806a3/_apis/git/repositories/0e3185be-f7ab-43e1-9595-7832fc0b4196/pullRequests/22411/attachments/screencap.png)

Will say "Frontend is opened." if the example frontend is selected.

- `common` contains the `CustomSystemUiPanel`
- `frontend` contains an `ExampleFrontend` using this new panel type
- `systemui` contains a very basic systemui that supports `CustomSystemUiPanel`, `TaskPanel` and `TaskProcessBarPanel`.