Merged PR 23822: IVI-7686: Disable trace event forwarding for release builds
The trace event forwarder is started in `Application.onCreate` to start it in all processes. However this causes the whole service framework to be initialized as part of `Application.onCreate` even if the process is not doing anything with `IviServices` or trace events. It also causes the `TraceEventService` to be started, which might never be used when the system does not have a trace event drain.

Allow the Application class to control whether trace event forwarding has to be started or not. The `ExampleApplication` class disables trace event forwarding for release build. The `IviApplication` keeps starting the trace event forwarding by default. The `IviApplication` cannot disable it based on the build type as we only publish release artifacts.