Merged PR 24161: IVI-7678: Introduce the service host concept
https://jira.tomtomgroup.com/browse/IVI-7678

Explain what a service host is and how it relates to an IVI service. This confused me a lot at the start. Link to https://developer.tomtom.com/tomtom-indigo/documentation/development/ivi-services (that article could also start from the beginning more and explain what a service host is and how it relates to an IVI service beyond the sentence "IVI service hosts can implement multiple IVI service interfaces."

https://jira.tomtomgroup.com/browse/IVI-7679
Start somewhere by saying that an IVI system is made up of IVI services which can provide a UI or another service with information