Merged PR 24853: IVI-7495: Quick fix for NavKit2 in Navigation page
This is a quick fix to remove NavKit2 from the portal page [Development > Platform Domains > Navigation](https://developer.tomtom.com/tomtom-indigo/documentation/development/platform-domains/navigation). For now we simply remove the "2" and leave "NavKit", to make the problem less obvious.
This page must be rewritten by someone with more knowledge of the Navigation feature in IndiGO.