Merged PR 26142: IVI-8447: Only publish Documentation if there are release notes
Changed the logic, so that the **Documentation** stage only runs if there are tickets with release notes.
If there are no release notes, Nexus publications and Github publication of the sources into the **beta** branch will still occur.

Removed also any text- and checkboxes in the pipeline menu related to release notes. As the release notes should only be picked up from jira tickets.