Merged PR 26676: IVI-8754: Do not publish empty oss scan reports
oss scan tasks were still succeeding in case of a stacktrace.
Let's still perform the scan tasks. Retry at max 3 times if needed. And publish the logs as pipeline artifacts for analysis. Whether or not the scans fail.

But do not create a group report md file in case of a failure. And do not publish this report  to artifactory, nor Github. Also let the pipeline fail in case of a stacktrace.