Merged PR 28584: IVI-9575: Write releases to Release readiness service
The stages of the SDK release pipeline have been updated to make more logical sense and also make the operations predictable. The DevPortal and CleanUp stage has been merged into the Publication as steps within a single job of publication. This ensures that the clean-up and retrigger can be done deterministically.

- Remove writing to 'Release Readiness Service' from the 'Publish' stage.
- Move writing to releases.json to the 'Publication' stage.
- Move the 'DevPortal' stage as a step within the 'Publication' Stage
- Move the 'Cleanup' stage as a step within the 'Publication Stage

- Updated 'Publication' stage has one job with following steps:
  - Write release to releases.json file & RRS
  - Generate portal documentation
  - Github publication
  - Cleanup

Additionally, the portal export step now invokes the python script directly instead of the earlier situation, where the gradle task would invoke the python script. This allowed us to use a lighter container for the job.  The gradle task is still part of the codebase.