Merged PR 28866: IVI-9739: Opt-in to debug groups and appsuite groups
The goal of this change is two fault:
- Only include debug tabs in debug builds by default. Same for services required for testing only.
- Be more explicit what which appsuite groups are included in the product.

Debug and test functionality, such as debug tabs, should not be included in release build that are used in production vehicles to reduce the attack surface. This change only adds the build time configuration groups `debugGroup` and `testGroup` in debug builds, by default.

To not break external projects, the `IviDefaultsGroupsSelectionConfigurator.includeDefaultGroups` still includes the build time configuration group `debugGroup` and is marked as deprecated. Usages can be replaced with two approaches:

1) Use `IviDefaultsGroupsSelectionConfigurator.includeDefaultPlatformGroups` and `IviDefaultsGroupsSelectionConfigurator.includeDefaultAppsuiteGroups`.

2) Use `IviDefaultsGroupsSelectionConfigurator.includeDefaultPlatformGroups` and explicitly opt-in to all desired appsuite groups with ``IviDefaultsGroupsSelectionConfigurator.include`.

As 2 is preferred for customers. As this makes it explicit which app suites are added and and explicit opt-in is required for customer to use a new appsuite group. For this reason our standalone indigo and the template app will use approach 2. Whereas examples and other demo products may prefer approach 1 to reduce maintenance.