Merged PR 28966: IVI-9806: Disable build-time check of API key by default for customers
Currently the API key build-time check is not disabled by default for customers.

This PR disables the API key build-time check by default.
And reverses the logic in the documentation, explaining that the advanced functionality is disabled by default and configuration steps can be skipped unless the customer needs this advanced functionality.